pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
name = "hr"
version = "0.0.0"
description = "Print horizontal rules."
authors = ["Martin Neufeld <m.neufeld946@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer-slim = "^0.12.3"

[tool.poetry.scripts]
hr = "hr.main:app"

[tool.poetry-dynamic-versioning]
enable = true
